Renamed reducer to dataStructureReducer and added combineReducer to manage several redusers together.
