+++ src/redux/reducers/index.ts
@@ -1,0 +1,8 @@
+import { combineReducers } from 'redux';
+import { dataStructureReducer } from './dataStructureReducer';
+
+export const rootReducer = combineReducers({
+  dataStructure: dataStructureReducer,
+});
+
+export type RootState = ReturnType<typeof rootReducer>;
